Build for Spark-3.1.2, scala 2.12 and hadoop-3.3.1, and python3.8+ support
Python 3.8 started requiring a type_ignores parameter in modules to the compile function, but the ast does not add that
by default thus breaking code submission. Not sure if simply adding an empty list (which is the default from the ast
parse function) is the correct long-term solution, but I believe it preserves the previous behavior.

Built with: mvn package -P spark-3.0 -DskipTests